+++ ntex-io/src/filter.rs
@@ -85,7 +85,7 @@ impl Filter for Base {
             } else if st.flags.is_stopping_filters() {
                 Poll::Ready(Readiness::Ready)
             } else if st.flags.is_read_paused_or_backpressure() {
-                // read buffer is fulled of is not processed by dispatcher yet
+                // read buffer is full or is not processed by dispatcher yet
                 Poll::Pending
             } else {
                 Poll::Ready(Readiness::Ready)